// src/modules/upload-minio/enums/upload.enum.ts
export enum EUploadError {
  NO_FILE_PROVIDED = 'NO_FILE_PROVIDED',
  FILE_TOO_LARGE = 'FILE_TOO_LARGE',
  INVALID_FILE_SIZE = 'INVALID_FILE_SIZE',
  INVALID_FILE_EXTENSION = 'INVALID_FILE_EXTENSION',
  INVALID_FILE_TYPE = 'INVALID_FILE_TYPE',
  UPLOAD_FAILED = 'UPLOAD_FAILED',
  MINIO_CONNECTION_ERROR = 'MINIO_CONNECTION_ERROR',
  BUCKET_NOT_FOUND = 'BUCKET_NOT_FOUND',
  FILE_NOT_FOUND = 'FILE_NOT_FOUND',
  DELETE_FAILED = 'DELETE_FAILED'
}

export enum EUploadStatus {
  PENDING = 'PENDING',
  UPLOADING = 'UPLOADING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  DELETED = 'DELETED'
}

export enum EAllowedFileType {
  IMAGE = 'image'
}

export enum EFileExtension {
  JPG = 'jpg',
  JPEG = 'jpeg',
  PNG = 'png'
}